完成增加和删除任务

+++ internal/timewheel/type.go
@@ -1,9 +1,8 @@
 package timewheel
 
-import "time"
 
 type chanInfo struct {
-	TaskId   uint16
+	TaskId   int
 	TaskPos  int
 	TaskTime int
 }
